Move `ScriptRunner` from `core-utils` to `common-utils`
Slightly modify the class so it can be used independently of ORT.
